requirements.txt
Django
djangorestframework
django-model-utils
psycopg2-binary
psycopg2
Pillow
django-filter
djangorestframework-simplejwt
drf-yasg
mock